--- src/pages/subscribe-page/subscribe.test.tsx
@@ -1,1 +1,0 @@
-export const nothing = 0;
